move managedlabel utils to new package
